requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-genai>=0.8.0
google-generativeai>=0.8.0
Pillow>=10.3.0
python-multipart>=0.0.9
